backend/requirements.txt
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
Markdown==3.2.2
django-model-utils==4.0.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
django-cors-headers==3.5.0
